requirements.txt
asgiref
Django
django-filter
djangorestframework
Pillow
psycopg2-binary
python-decouple
pytz
sqlparse
